+++ app/Http/Controllers/InfoCaracLevadurasController.php
@@ -48,7 +48,7 @@ class InfoCaracLevadurasController extends Controller
                 $this->crearSeguimiento("Agregó un Tipo de Textura en Levaduras: "
                     . $tipo->nombre);
                 break;
-            case "metodo_conser":
+            case "tipo_metodo":
                 $rules = [
                     'nombre' => 'bail|required|unique:tipo_metodo_conservacion_levaduras,nombre'
                 ];
@@ -96,7 +96,7 @@ class InfoCaracLevadurasController extends Controller
                 $this->crearSeguimiento("Editó un Tipo de Textura en Levaduras: "
                     . $tipo1->nombre);
                 break;
-            case "metodo_conser":
+            case "tipo_metodo":
                 $tipo1 = TipoMetodoConservacionLevadura::find($id);
                 $tipo2 = TipoMetodoConservacionLevadura::where('nombre', $request->nombre)->first();
                 if (!is_null($tipo2)) {
@@ -119,24 +119,36 @@ class InfoCaracLevadurasController extends Controller
         switch ($request->tipo) {
             case "color":
                 $tipo = ColorLevadura::find($id);
-                broadcast(new LevadurasInfoEvent($tipo, $request->tipo, 'eliminar'))->toOthers();
-                $tipo->delete();
-                $this->crearSeguimiento("Eliminó un Tipo de Color en Levaduras: "
-                    . $tipo->nombre);
+                if ($this->validarEliminar($tipo, 'macro')) {
+                    return 'macro';
+                } else {
+                    broadcast(new LevadurasInfoEvent($tipo, $request->tipo, 'eliminar'))->toOthers();
+                    $tipo->delete();
+                    $this->crearSeguimiento("Eliminó un Tipo de Color en Levaduras: "
+                        . $tipo->nombre);
+                }
                 break;
             case "textura":
                 $tipo = TexturaLevadura::find($id);
-                broadcast(new LevadurasInfoEvent($tipo, $request->tipo, 'eliminar'))->toOthers();
-                $tipo->delete();
-                $this->crearSeguimiento("Eliminó un Tipo de Textura en Levaduras: "
-                    . $tipo->nombre);
+                if ($this->validarEliminar($tipo, 'macro')) {
+                    return 'macro';
+                } else {
+                    broadcast(new LevadurasInfoEvent($tipo, $request->tipo, 'eliminar'))->toOthers();
+                    $tipo->delete();
+                    $this->crearSeguimiento("Eliminó un Tipo de Textura en Levaduras: "
+                        . $tipo->nombre);
+                }
                 break;
-            case "metodo_conser":
+            case "tipo_metodo":
                 $tipo = TipoMetodoConservacionLevadura::find($id);
-                broadcast(new LevadurasInfoEvent($tipo, $request->tipo, 'eliminar'))->toOthers();
-                $tipo->delete();
-                $this->crearSeguimiento("Eliminó un Tipo de Metodo de Conservación en Levaduras: "
-                    . $tipo->nombre);
+                if ($this->validarEliminar($tipo, 'metodo')) {
+                    return 'metodo';
+                } else {
+                    broadcast(new LevadurasInfoEvent($tipo, $request->tipo, 'eliminar'))->toOthers();
+                    $tipo->delete();
+                    $this->crearSeguimiento("Eliminó un Tipo de Metodo de Conservación en Levaduras: "
+                        . $tipo->nombre);
+                }
                 break;
         }
 
@@ -151,5 +163,27 @@ class InfoCaracLevadurasController extends Controller
         $seguimiento->tipo_user = Auth::user()->tipouser->nombre;
         $seguimiento->accion = $accion;
         $seguimiento->save();
+    }
+
+    public function validarEliminar($tipoCaract, $tipo)
+    {
+        $res = false;
+        switch ($tipo) {
+            case "macro":
+                if (
+                    count($tipoCaract->caractMacros) > 0
+                ) {
+                    $res = true;
+                }
+                break;
+            case "metodo":
+                if (
+                    count($tipoCaract->metodosConservacion) > 0
+                ) {
+                    $res = true;
+                }
+                break;
+        }
+        return $res;
     }
 }
